% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_mirna.R
\name{add_mirna}
\alias{add_mirna}
\title{Add miRNA target data to an edit_table object}
\usage{
add_mirna(this, mirna_file)
}
\arguments{
\item{this}{a edit_table object}

\item{mirna_file}{character string naming formatted miranda ouptut file}
}
\value{
a newly formatted edit_table object with miRNA target information
}
\description{
Add miRNA target data to an edit_table object
}

