% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_repeatmask.R
\name{add_repeatmask}
\alias{add_repeatmask}
\title{Add repeatmasker hits to an edit_table object}
\usage{
add_repeatmask(this, rm_file)
}
\arguments{
\item{this}{a edit_table object}

\item{rm_file}{character string naming repeatmasker file}
}
\value{
a newly formatted edit_table object with repeatmasker information
}
\description{
Add repeatmasker hits to an edit_table object
}

