% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_vep.R
\name{add_vep}
\alias{add_vep}
\title{Adds Variant Effect Predictor information to an edit_table object}
\usage{
add_vep(this, vep_file)
}
\arguments{
\item{this}{an edit_table object}

\item{vep_file}{character giving path to txt file with VEP results}
}
\value{
a new edit_table object with added $VEP field
}
\description{
Adds Variant Effect Predictor information to an edit_table object
}

