% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/edit_scan.R
\name{edit_scan}
\alias{edit_scan}
\title{Performs a scan of all putative RNA editing sites from vcf files}
\usage{
edit_scan(plus_file, minus_file, names = c("DNA", "RNA"), qual = 10,
  ex.indel = TRUE, geno.dp = 10, geno.hom = 95, edit.dp = 5,
  summary = TRUE)
}
\arguments{
\item{plus_file}{}

\item{minus_file}{}
}
\value{
an edit_summary object
}
\description{
Requires two VCF files as input. One of the files requires RNA sequencing data
 from PLUS strand alignments, the other from MINUS strand alignments.
 Each file needs to have two samples. One sample will always be from genomic DNA.
 The other will be from a RNA sequencing data from a tissue of interest.
}

