% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/edit_summary.R
\name{edit_summary}
\alias{edit_summary}
\title{Scans vcf class for evidence of RNA editing}
\usage{
edit_summary(obj, qual = 10, ex.indel = TRUE, geno.dp = 10,
  geno.hom = 95, edit.dp = 5, summary = TRUE)
}
\arguments{
\item{obj}{A vcf object}

\item{qual}{An integer specifiying the minimum variant QUAL}

\item{ex.indel}{logical indicating whether to exclude indels from the scan}

\item{geno.dp}{integer specifying the minimum genotype depth}

\item{geno.hom}{integer ranging from 0 to 1 specifiying the proportion of homozygosity
the genotype must exhibit}

\item{edit.dp}{integer specifying the minimum depth required for evidence of
RNA editing}

\item{summary}{logical indicating whether a summary of found edits is returned
or a subsetted vcf object}
}
\value{
Either a subsetted vcf object, only containing probable edited loci
 or an object of class "edit_summary"
}
\description{
Scans vcf class for evidence of RNA editing
}

