% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ErnstEditTools.R
\docType{package}
\name{ErnstEditTools}
\alias{ErnstEditTools}
\alias{ErnstEditTools-package}
\title{ErnstEditTools: Detect and analyze RNA editing data from vcf input}
\description{
Example data was generated by grabbing the first 5000 lines of each
raw .vcf file (after the header)
}
\details{
Note that this package is still in development. Compiled code is
intended to be incorporated to dramatically improve performance.
}

