% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/facet_plot.R
\name{facet_plot}
\alias{facet_plot}
\title{Turns two or more plots of edit_tables into a facet plot}
\usage{
facet_plot(..., names = NULL, text_size = 20, percent_size = 10,
  group = "Tissue")
}
\arguments{
\item{...}{any number data.frames resulting from plot.edit_table(plot = FALSE)}

\item{names}{character vector for each sample provided in}

\item{text_size}{numeric changing the appearence of text in the plot}

\item{group}{character providing the 'fill' aesthetic argument fo each facet}
}
\value{
a gg object
}
\description{
Turns two or more plots of edit_tables into a facet plot
}

