% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/find_edits.R
\name{find_edits}
\alias{find_edits}
\title{Processes VCF files in search for candidate RNA editing events}
\usage{
find_edits(file_plus, file_minus = NULL, names = character(), qual = 10,
  ex_indel = TRUE, geno_dp = 10, geno_hom = 95, edit_dp = 5)
}
\arguments{
\item{file_plus}{input filename for VCF file 1.}

\item{file_minus}{input filename for VCF file 2. If only one VCF file provided, it is assumed
that the RNA variants in that file all come from plus strand transcripts.}

\item{names}{A character vector}

\item{qual}{An integer specifiying the minimum variant QUAL}

\item{ex_indel}{logical indicating whether to exclude indels from the scan}

\item{geno_dp}{integer specifying the minimum genotype depth}

\item{geno_hom}{integer ranging from 0 to 1 specifiying the proportion of homozygosity
the genotype must exhibit}

\item{edit_dp}{integer specifying the minimum depth required for evidence of
RNA editing}

\item{summary}{logical indicating whether a summary of found edits is returned
or a subsetted vcf object}
}
\value{
an edit_summary object
}
\description{
Must supply two files - one from RNA seq alignments that come from
 plus strand transcripts (VCF file 1). The other from minus strand transcripts.
 (VCF file 2).
}
\details{
Each vcf file requires a genomic DNA sample (the first sample listed in the vcf file),
 along with any number of RNA samples from various tissues.
}

