% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/merge_strands.R
\name{merge_strands}
\alias{merge_strands}
\title{Merges edit_summary objects from different strands}
\usage{
merge_strands(plus_obj, minus_obj)
}
\arguments{
\item{plus_obj}{an edit_summary object - generated from RNA sequencing reads
that mapped to the PLUS strand}

\item{minus_obj}{an edit_summary object - generaged from RNA sequencing reads
that mapped ot the MINUS strand}
}
\value{
a new edit_summary object that is the result of properly merging
 both plus_obj and minus_obj
}
\description{
Merges edit_summary objects from different strands
}

