% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mq_filter.R
\name{mq_filter}
\alias{mq_filter}
\title{Filters potential RNA editing sites by applying filtering on average mapping quality (MQ)}
\usage{
mq_filter(this)
}
\arguments{
\item{this}{an edit_table object}
}
\description{
Calculates P(X >= x | N, p(e)) where x is the number of reads which provide support for
editing for sample i at site j, N is the total depth for sample i at site j,
and p(e) is the probability of error for site j, derived from the average mapping quality.
}

