% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_edits.R
\name{plot_edits}
\alias{plot_edits}
\title{A plotting tool to visualize DNA RNA mismatch counts from an edit_frame}
\usage{
plot_edits(this)
}
\description{
A plotting tool to visualize DNA RNA mismatch counts from an edit_frame
}

