% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vcf.R
\name{read_vcf}
\alias{read_vcf}
\title{Read in .vcf files}
\usage{
read_vcf(filename, names = NULL)
}
\arguments{
\item{filename}{The filename of the .vcf file}

\item{names}{character vector of names that will be used to reference each sample. 
Specify in the order that the samples appear in the file.}
}
\value{
An object of class vcf
}
\description{
Initializes a vcf object for use in editTools methods
Currently, only works with 2-sample VCF files
}

