% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{samples}
\alias{samples}
\title{Getter for Samples field}
\usage{
samples(obj, n)
}
\arguments{
\item{obj}{a vcf object}

\item{n}{either a numeric or character identifying which
sample to extract}
}
\value{
Samples field (a list of samples) or if n provided,
 a data.frame with information for a single sample.
}
\description{
Getter for Samples field
}

