% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{[.vcf}
\alias{[.vcf}
\title{Subsetting method for vcf class}
\usage{
\method{[}{vcf}(obj, p, n)
}
\arguments{
\item{obj}{a vcf obect}

\item{p}{numeric to filter SNPs}

\item{n}{numeric to filter Samples}
}
\value{
a subsetted vcf object
}
\description{
Subsetting method for vcf class
}

