% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset.edit_table}
\alias{subset.edit_table}
\title{Subset editing candidates from an edit_table object}
\usage{
\method{subset}{edit_table}(this, dna = NULL, rna = NULL, tissue = NULL)
}
\arguments{
\item{this}{an edit_table object}

\item{dna}{character indicating desired DNA call}

\item{rna}{character indicating desired RNA call}

\item{tissue}{character indicating desired tissue sample}
}
\value{
a subsetted edit_table
}
\description{
WORK IN PROGRESS
}
\details{
Can subset based on such criteria as tissue type or DNA call

If subsetted both by DNA and RNA call (mismatch type) then a
 data.frame is returned rather than an edit_table object
}

