% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write_vep.R
\name{write_vep}
\alias{write_vep}
\title{Writes a txt file suitable for Variant Effect Predictor software}
\usage{
write_vep(this, file = "edits.vep")
}
\arguments{
\item{this}{an edit_table object}

\item{file}{character giving desired output file}
}
\value{
none
}
\description{
Writes a txt file suitable for Variant Effect Predictor software
}

