% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.edit_table.R
\name{plot.edit_table}
\alias{plot.edit_table}
\title{A plotting tool to visualize data from an edit_frame object}
\usage{
\method{plot}{edit_table}(this, field = "AllSites", event = "Mismatch",
  mismatch = "all", group = "Tissue", n = NULL, text_size = 20,
  perc_size = 10, perc_round = 0, plot = TRUE,
  x_label = "Mismatch type", y_label = "Number of events",
  y_range = "auto", denom_all = FALSE, legend = TRUE)
}
\arguments{
\item{this}{an edit_table object}

\item{field}{character naming member of edit_table to plot. A barplot is produced
for members $AllSites, $RepSites, $mirnaTargetSites, and $VEP. A Venn Diagram
is produced for member $Tissues}

\item{event}{character naming which column from field to count}

\item{mismatch}{character indicating which type of mismatch to consider, such as "AtoG".
By default, considers all mismatch types}

\item{group}{character selecting field to group (color code) by.}

\item{n}{integer limiting number of events to n. Any number of events beyond
n will be grouped into "other" category}

\item{text_size}{numeric giving the size of axis and legend labels.}

\item{perc_size}{numeric giving the size of percent labels above bars.}

\item{perc_round}{numeric giving decimal places that percent labels round to.}

\item{plot}{logical where TRUE returns a gg plot object. False returns the modified data.frame from
which the gg object is constructed from. False can be used to combine datasets for "facet plotting".}

\item{x_label}{character providing label for x axis}

\item{y_label}{character providing label for y axis}

\item{y_range}{numeric vector providing min and max of y axis range}

\item{denom_all}{logical. If true, percentage labels will always be
out of the total number of edits found. If false, only the number of edits
in field will be considered.}

\item{legend}{logical. If false, any legend that would otherwise appear is turned off.}
}
\description{
A plotting tool to visualize data from an edit_frame object
}

