% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tissue_plot.R
\name{tissue_plot}
\alias{tissue_plot}
\title{Plotting tool to produce venn diagrams showing off tissue specific mismatch counts}
\usage{
tissue_plot(this, field = "AllSites", fill_colors = "green",
  mismatch = "all")
}
\arguments{
\item{this}{an edit_table object}

\item{field}{character naming edit_table field (either $AllSites,
$RepSites, or $mirnaTargetSites) from which to obtain counts from}

\item{fill_colors}{character vector specifying colors for each tissue}
}
\description{
Plotting tool to produce venn diagrams showing off tissue specific mismatch counts
}

